refactor: add verify message status modal
