--- src/domains/wallets/components/MessageStatus/index.ts
@@ -1,1 +1,0 @@
-export * from "./MessageStatus";
