--- src/domains/wallets/components/MessageStatus/templates/VerifyMessage/index.ts
@@ -1,1 +1,0 @@
-export * from "./VerifyMessage";
