+++ src/domains/wallets/components/VerifyMessageStatus/index.ts
@@ -1,0 +1,1 @@
+export * from "./VerifyMessageStatus";
